// frontend/src/app/model/employee.ts
export interface Employee {
    id: string;
    first_name: string;
    last_name: string;
    email: string;
    gender: string;
    designation: string;
    salary: number;
    date_of_joining: string;
    department: string;
    created_at: string;
    updated_at: string;
  }
  
  export interface NewEmployee {
    first_name: string;
    last_name: string;
    email: string;
    gender: string;
    designation: string;
    salary: number;
    date_of_joining: string;
    department: string;
  }
  